ascii-art-converter-python/requirements.txt
Pillow==5.2.0
numpy==1.15.1
